First Edition of Vending Machine
Forum created. At the moment we can get money, choose a single element, take it and see the decline in its number. We can also get warnings of insufficient money. When we enter an incorrect value, we receive a warning that we do not entered a number. We can select other elements in the next version and we will be able to see the decrease in the numbers.